requirements.txt
streamlit
groq
gitpython
python-dotenv
pandas
requests
